FakeRoot: implement copy
This adds a copy method to FakeRoot, that copies a tree of files/directories/symlinks.
The copying is done by python run under fakeroot, to allow creating all files with
the correct permission and owner/group recorded in the fakeroot state.
It uses hardlinks to copy files, so metadata is inherited for all files and it is cheap.

TODO: Add documentation comment
TODO: Add fallback to copy instead of hardlink, if hardlink is not possible