MBR: fix compatibility with libparted >= 3.6
With parted 3.6 a check was implemented that allows setting the LBA flag
only for msdos partitions that support it (i.e. all fat types, DOS_EXT and EXT_LBA).

To fix the issue only set the flag when it is supported.

Aditionally extended partitions are correctly set to type 0x0f (EXT_LBA),
which was not done before. This also fixes it for older parted releases